% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster2.R
\name{xy_rgb}
\alias{xy_rgb}
\title{Return data frame of RGB color for pixels from `magick` image data}
\usage{
xy_rgb(x, y, rgbm = rgb, x1 = x1, xstp = 0.5, y1 = y1, ystp = 0.5)
}
\arguments{
\item{x}{x coordinates}

\item{y}{y coordinates}

\item{rgbm}{matrix of RGB color generated from `magick`}

\item{x1}{starting point of x}

\item{xstp}{step of x sampling interval}

\item{y1}{starting point of y}

\item{ystp}{xstp step of y sampling interval}
}
\value{
data frame of RGB color for pixels
}
\description{
Return data frame of RGB color for pixels from `magick` image data
}
\examples{
img<-magick::image_read("07KE0024_office.tif")
plot(img)
rgb <- as.integer(img[[1]])
x1<-3780;x2<-4000;y1<--1170;y2<- -950　###   平面直角座標系の位山宿舎の範囲
xstp <- 0.5 ; ystp <- 0.5　#画素
load("KuraiyamaOfficeDSM_xyz.RData")
dsm<-xyz
rgb_ <- xy_rgb(dsm$x,dsm$y,rgb,x1,0.5,y1,0.5)
open3d()
points3d(dsm,col=rgb(rgb_,maxColorValue = 255))
}
